+++ packages/dts-plugin/src/server/message/Action/AddDynamicRemote.ts
@@ -1,8 +1,6 @@
-import { UpdateMode } from '../../../core/index';
 import { RemoteInfo } from '../../../core/interfaces/HostOptions';
 import { BaseContext } from '../../types';
 import { Action, ActionKind } from './Action';
-import { UpdateKind } from './Update';
 
 // host consumes provider , the action will notify host fetch provider types
 export interface AddDynamicRemotePayload extends BaseContext {