+++ packages/dts-plugin/src/server/message/Action/index.ts
@@ -9,4 +9,4 @@ export * from './NotifyWebClient';
 export * from './UpdatePublisher';
 export * from './FetchTypes';
 export * from './AddDynamicRemote';
-export { UpdateMode } from '../../../core/constant';
+export { UpdateMode } from '../../../server/constant';