+++ packages/bridge/bridge-react-webpack-plugin/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @module-federation/bridge-react-webpack-plugin
 
+## 0.3.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.1
+
+## 0.3.0
+
+### Patch Changes
+
+- Updated dependencies [fa37cc4]
+  - @module-federation/sdk@0.3.0
+
+## 0.2.8
+
+### Patch Changes
+
+- @module-federation/sdk@0.2.8
+
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [b00ef13]
+  - @module-federation/sdk@0.2.7
+
 ## 0.2.6
 
 ### Patch Changes