+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @module-federation/esbuild
 
+## 0.0.9
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.1
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [fa37cc4]
+  - @module-federation/sdk@0.3.0
+
+## 0.0.7
+
+### Patch Changes
+
+- @module-federation/sdk@0.2.8
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [b00ef13]
+  - @module-federation/sdk@0.2.7
+
 ## 0.0.5
 
 ### Patch Changes